11239: Simplify report rendering
