+++ app/views/reports/_main.html.erb
@@ -1,8 +1,6 @@
 <div class="tip report_info"></div>
 <div class="report_body"></div>
 
-<% if @report_data %>
-  <%= javascript_doc_ready do %>
-    ELMO.app.report_controller = new ELMO.Report.ReportController(<%=json @report_data %>);
-  <% end %>
+<%= javascript_doc_ready do %>
+  new ELMO.Report.ReportController(<%=json @report_data %>);
 <% end %>
