+++ app/views/reports/show.html.erb
@@ -5,7 +5,7 @@
 </div>
 <div class="report_view">
   <div class="report-main">
-    <%= render('main') %>
+    <%= render("main") %>
   </div>
 </div>
 
@@ -312,7 +312,3 @@
     </div><!-- /.modal-dialog -->
   </form>
 </div><!-- /.modal -->
-
-<%= javascript_doc_ready do %>
-  ELMO.app.report_controller.run_report();
-<% end unless @report.new_record? || flash[:edit_mode] %>
